Move the registry tests into test_s3.py
Keep the test file layout mirroring the source layout instead of adding
a separate test_registry.py.
